+++ tidy3d/components/grid/grid.py
@@ -10,7 +10,7 @@ import pydantic.v1 as pd
 from ...exceptions import SetupError
 from ..base import Tidy3dBaseModel
 from ..data.data_array import DataArray, ScalarFieldDataArray, SpatialDataArray
-from ..data.dataset import UnstructuredGridDataset, UnstructuredGridDatasetType
+from ..data.utils import UnstructuredGridDataset, UnstructuredGridDatasetType
 from ..geometry.base import Box
 from ..types import ArrayFloat1D, Axis, InterpMethod, Literal
