+++ tidy3d/components/parameter_perturbation.py
@@ -18,13 +18,13 @@ from ..exceptions import DataError
 from ..log import log
 from .base import Tidy3dBaseModel, cached_property
 from .data.data_array import ChargeDataArray, HeatDataArray, IndexedDataArray, SpatialDataArray
-from .data.dataset import (
+from .data.utils import (
     CustomSpatialDataType,
-    UnstructuredGridDataset,
     _check_same_coordinates,
     _get_numpy_array,
     _zeros_like,
 )
+from .data.unstructured.base import UnstructuredGridDataset
 
 """ Generic perturbation classes """
